blackjack logic
