+++ frontend/src/components/blackjack/blackjack_container.js
@@ -1,0 +1,8 @@
+import { connect } from 'react-redux';
+import Blackjack from './blackjack';
+
+const mapStateToProps = state => ({
+    currentUser: state.session.user
+});
+
+export default connect(mapStateToProps)(Blackjack);
