+++ frontend/src/components/blackjack/card.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+class Card extends React.Component {
+    
+
+    render() {
+        return (
+            <div>
+                hello
+            </div>
+        );
+  }
+}
+
+export default Card;
